v 0.22; completed update_one and used it on chase;

+++ django2wrap/models.py
@@ -92,7 +92,7 @@ class Case(models.Model):
     in_support_sla = models.BooleanField(default=True) # actual
     in_response_sla = models.BooleanField(default=True) # actual
     support_sla = models.PositiveIntegerField() # sla # in hrs
-    response_sla = models.PositiveIntegerField() # sla # in hrs
+    response_sla = models.FloatField() # sla # in hrs
     support_time = models.FloatField(null=True) # casetimes ## resolution time in hrs
     response_time = models.FloatField(null=True) # casetimes ## response time in hrs
     raw = models.TextField(null=True) # this should be added once on stable DB
